+++ frontend/src/components/admin/payrollNavbar.js
@@ -1,27 +1,38 @@
 import React, { Component } from 'react';
-import MuiAppBar from '@mui/material/AppBar';
 import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
+import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import List from '@mui/material/List';
 import Typography from '@mui/material/Typography';
+import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
-import Link from '@mui/material/Link';
+import Paper from '@mui/material/Paper';
 import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import { mainListItems } from './listSessionOptions';
-import logo from '../../../images/logo.png';
+import { mainListItems} from '../../pages/admin/Payroll/listItems';
+import Spendings from '../../pages/admin/Payroll/Spendings';
+import MonthlySpendings from '../../pages/admin/Payroll/MonthlySpendings';
+import YearlySpendings from '../../pages/admin/Payroll/YearlySpendings';
+import TutorSnapshot from '../../pages/admin/Payroll/tutorSnapshot';
+import logo from '../../images/logo.png';
 import axios from 'axios';
+import TutorsList from '../../pages/admin/Payroll/tutors';
+import PayRate from '../../pages/admin/Payroll/payRate';
 import LogoutIcon from '@mui/icons-material/Logout';
 import { useNavigate } from 'react-router-dom';
-import Sessions from './sessions';
-import AddSession from './addSession';
+import Link from '@mui/material/Link';
 
-// Component to display copyright information
+//this is the navbar
+
+import {
+  createStyles,
+} from '@material-ui/core';
+
 function Copyright(props) {
   return (
     <Typography variant="body2" color="text.secondary" align="center" {...props}>
@@ -35,19 +46,19 @@ function Copyright(props) {
   );
 }
 
-// Width of the drawer
-const drawerWidth = 200;
 
-// Creating a default theme for the application
-const defaultTheme = createTheme({
-  components: {
-    MuiPaper: {
-      styleOverrides: {
-        root: {
-          // Custom styling for MuiPaper component
-        },
-      },
-    },
+const drawerWidth = 240;
+const styles = createStyles({
+  container: {
+    padding: '20px',
+  },
+  tableContainer: {
+    marginTop: '20px',
+  },
+  pagination: {
+    marginTop: '20px',
+    display: 'flex',
+    justifyContent: 'center',
   },
 });
 
@@ -75,7 +86,7 @@ const AppBar = styled(MuiAppBar, {
   }),
 }));
 
-// Custom styled Drawer component using MuiDrawer
+
 const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
   ({ theme, open }) => ({
     '& .MuiDrawer-paper': {
@@ -102,35 +113,80 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
   }),
 );
 
-class SessionDashbaord extends Component {
+// TODO remove, this demo shouldn't need to reset the theme.
+const defaultTheme = createTheme();
+
+class Payroll extends Component {
   constructor() {
     super();
     this.state = {
-      showSessions: true,
-      addSession: false,
-      showReports: false,
+      token: '',
+      page: 1,
+      users: [],
+      pages: 0,
+      showTutors: false,
+      showPayRate: false,
       loading: false,
     };
   }
 
-    // Toggle the drawer open/close state
-    toggleDrawer = () => {
-      this.setState((prevState) => ({ open: !prevState.open }));
-    };
+  handleTutorsClick = () => {
+    this.setState({ showTutors: true, showPayRate: false, showCharts: false });
+  };
   
-  // Event handler for the register button click
-  handleSessionsClick = () => {
-    this.setState({ showSessions: true, addSession: false, showReports: false });
+  handlePayRateClick = () => {
+    this.setState({ showPayRate: true, showTutors: false, showCharts: false });
+  };
+  handleDashboardClick = () => {
+    this.setState({ showTutors: false, showPayRate: false, showCharts: false });
+  };
+  
+
+  componentDidMount = () => {
+    let token = localStorage.getItem('token');
+    if (!token) {
+      this.props.navigate('/login');
+    } else {
+      this.setState({ token: token }, () => {
+        this.getSession();
+
+      });
+    }
   };
 
-  // Event handler for the accounts button click
-  handleAddSessionClick = () => {
-    this.setState({ showSessions: false, addSession: true, showReports: false });
+  getSession = () => {
+    this.setState({ loading: true });
+
+    let data = '?';
+    data = `${data}page=${this.state.page}`;
+    
+    axios
+      .get(`http://localhost:2000/get-users${data}`, {
+        headers: {
+          token: this.state.token,
+        },
+      })
+      .then((res) => {
+        this.setState({ loading: false, users: res.data.users, pages: res.data.pages });
+      })
+      .catch((err) => {
+        this.setState({ loading: false, users: [], pages: 0 });
+      });
   };
 
-  // Event handler for the charts button click
-  handleReportsClick = () => {
-    this.setState({ showSessions: false, addSession: false, showReports: true });
+  pageChange = (e, page) => {
+    this.setState({ page: page }, () => {
+      this.getSession();
+    });
+  };
+
+  handleChartsClick = () => {
+    this.setState({ showTutors: false, showPayRate: false, showCharts: true });
+  };
+
+
+  toggleDrawer = () => {
+    this.setState((prevState) => ({ open: !prevState.open }));
   };
 
   LogOutButton = () => {
@@ -163,9 +219,9 @@ class SessionDashbaord extends Component {
       </IconButton>
     );
   };
-  
+
   render() {
-    const { open, showSessions, addSession, showReports } = this.state;
+    const { open, showTutors, showPayRate, showCharts } = this.state;
     const { LogOutButton, LogoButton } = this;
   
     return (
@@ -190,20 +246,23 @@ class SessionDashbaord extends Component {
               >
                 <MenuIcon />
               </IconButton>
-              <Typography
+                 <Typography
                 component="h1"
                 variant="h5"
                 color="white"
                 noWrap
                 sx={{ flexGrow: 1, fontWeight: 'bold', marginLeft: '150px' }} // Adjust this value to your needs
               >
-                Sessions Dashboard
+                Payroll
               </Typography>
               <div>
-                <LogoButton />
+              <LogoButton />
+
                 <span style={{ marginRight: '10px' }}></span> {/* Add space between the two icons */}
+
                 <LogOutButton />
               </div>
+            
             </Toolbar>
           </AppBar>
           <Drawer variant="permanent" open={open}>
@@ -219,16 +278,19 @@ class SessionDashbaord extends Component {
                 <ChevronLeftIcon />
               </IconButton>
             </Toolbar>
+            <Divider />
             <List component="nav">
-              {/* Render the list Session Options */}
-              {mainListItems(this.handleSessionsClick, this.handleAddSessionClick, this.handleReportsClick)}
+              {mainListItems(this.handleTutorsClick, this.handlePayRateClick, this.handleDashboardClick, this.handleChartsClick)}
+              <Divider sx={{ my: 1 }} />
             </List>
           </Drawer>
           <Box
             component="main"
             sx={{
               backgroundColor: (theme) =>
-                theme.palette.mode === 'light' ? theme.palette.grey[100] : theme.palette.grey[900],
+                theme.palette.mode === 'light'
+                  ? theme.palette.grey[100]
+                  : theme.palette.grey[900],
               flexGrow: 1,
               height: '100vh',
               overflow: 'auto',
@@ -236,23 +298,71 @@ class SessionDashbaord extends Component {
           >
             <Toolbar />
             <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-              {/* Conditionally render the components based on the state */}
-              {showSessions && <Sessions/>}
-              {addSession && <AddSession/>}
-              {showReports && (
-                <iframe
-                style={{
-                  background: '#F1F5F4',
-                  border: 'none',
-                  borderRadius: '2px',
-                  boxShadow: '0 2px 10px 0 rgba(70, 76, 79, .2)',
-                  width: '80vw',
-                  height: '100vh',
-                }}
-                src="https://charts.mongodb.com/charts-project-0-oyiwi/embed/dashboards?id=a6bc5c09-215f-4d3d-942e-fc32036d064f&theme=light&autoRefresh=true&maxDataAge=300&showTitleAndDesc=false&scalingWidth=scale&scalingHeight=scale"
-              />
-              )}
-              {/* Render copyright component */}
+              {showTutors ? <TutorsList /> : null}
+              {showPayRate ? <PayRate /> : null}
+              {showCharts ? (
+                <iframe 
+                  style={{
+                    background: '#F1F5F4',
+                    border: 'none',
+                    borderRadius: '2px',
+                    boxShadow: '0 2px 10px 0 rgba(70, 76, 79, .2)',
+                    width: '80vw',
+                    height: '100vh'
+                  }}
+                  src="https://charts.mongodb.com/charts-project-0-oyiwi/embed/dashboards?id=c075600e-06f7-4cd2-b9f3-a0bc9ba537c0&theme=light&autoRefresh=true&maxDataAge=3600&showTitleAndDesc=false&scalingWidth=scale&scalingHeight=scale"
+                />
+              ) : null}
+              {!showTutors && !showPayRate && !showCharts ? (
+                <Grid container spacing={3} justify="center">
+                  {/* Recent Deposits */}
+                  <Grid item xs={12} sm={4} md={4} lg={4}>
+                    <Paper
+                      sx={{
+                        p: 2,
+                        display: 'flex',
+                        flexDirection: 'column',
+                        height: 240,
+                      }}
+                    >
+                      <Spendings />
+                    </Paper>
+                  </Grid>
+                  
+                  {/* Recent Deposits */}
+                  <Grid item xs={12} sm={4} md={4} lg={4}>
+                    <Paper
+                      sx={{
+                        p: 2,
+                        display: 'flex',
+                        flexDirection: 'column',
+                        height: 240,
+                      }}
+                    >
+                      <MonthlySpendings />
+                    </Paper>
+                  </Grid>
+                  {/* Recent Deposits */}
+                  <Grid item xs={12} sm={4} md={4} lg={4}>
+                    <Paper
+                      sx={{
+                        p: 2,
+                        display: 'flex',
+                        flexDirection: 'column',
+                        height: 240,
+                      }}
+                    >
+                      <YearlySpendings />
+                    </Paper>
+                  </Grid>
+                  {/* Recent Sessions */}
+                  <Grid item xs={12}>
+                    <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+                      <TutorSnapshot />
+                    </Paper>
+                  </Grid>
+                </Grid>
+              ) : null}
               <Copyright sx={{ pt: 4 }} />
             </Container>
           </Box>
@@ -261,7 +371,9 @@ class SessionDashbaord extends Component {
     );
   }
   
+
+  
   
 }
 
-export default  (SessionDashbaord);
+export default Payroll;
