+++ frontend/src/components/shared/addSession.js
