+++ frontend/src/components/tutor/tutorSessionsNavbar.js
@@ -1,38 +1,27 @@
 import React, { Component } from 'react';
+import MuiAppBar from '@mui/material/AppBar';
 import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import List from '@mui/material/List';
 import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
+import Link from '@mui/material/Link';
 import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import { mainListItems} from './listItems';
-import Spendings from './Spendings';
-import MonthlySpendings from './MonthlySpendings';
-import YearlySpendings from './YearlySpendings';
-import TutorSnapshot from './tutorSnapshot';
-import logo from '../../../images/logo.png';
+import { mainListItems } from '../../pages/admin/AdminSessionsDashboard/listSessionOptions';
+import logo from '../../images/logo.png';
 import axios from 'axios';
-import TutorsList from './tutors';
-import PayRate from './payRate';
 import LogoutIcon from '@mui/icons-material/Logout';
 import { useNavigate } from 'react-router-dom';
-import Link from '@mui/material/Link';
+import Sessions from '../../pages/tutor/TutorSessionsDashboard/sessions';
+import AddSession from '../shared/addSession';
 
-//this is the navbar
-
-import {
-  createStyles,
-} from '@material-ui/core';
-
+// Component to display copyright information
 function Copyright(props) {
   return (
     <Typography variant="body2" color="text.secondary" align="center" {...props}>
@@ -46,19 +35,19 @@ function Copyright(props) {
   );
 }
 
+// Width of the drawer
+const drawerWidth = 200;
 
-const drawerWidth = 240;
-const styles = createStyles({
-  container: {
-    padding: '20px',
-  },
-  tableContainer: {
-    marginTop: '20px',
-  },
-  pagination: {
-    marginTop: '20px',
-    display: 'flex',
-    justifyContent: 'center',
+// Creating a default theme for the application
+const defaultTheme = createTheme({
+  components: {
+    MuiPaper: {
+      styleOverrides: {
+        root: {
+          // Custom styling for MuiPaper component
+        },
+      },
+    },
   },
 });
 
@@ -86,7 +75,7 @@ const AppBar = styled(MuiAppBar, {
   }),
 }));
 
-
+// Custom styled Drawer component using MuiDrawer
 const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
   ({ theme, open }) => ({
     '& .MuiDrawer-paper': {
@@ -113,80 +102,35 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
   }),
 );
 
-// TODO remove, this demo shouldn't need to reset the theme.
-const defaultTheme = createTheme();
-
-class Payroll extends Component {
+class SessionDashbaord extends Component {
   constructor() {
     super();
     this.state = {
-      token: '',
-      page: 1,
-      users: [],
-      pages: 0,
-      showTutors: false,
-      showPayRate: false,
+      showSessions: true,
+      addSession: false,
+      showReports: false,
       loading: false,
     };
   }
 
-  handleTutorsClick = () => {
-    this.setState({ showTutors: true, showPayRate: false, showCharts: false });
-  };
+    // Toggle the drawer open/close state
+    toggleDrawer = () => {
+      this.setState((prevState) => ({ open: !prevState.open }));
+    };
   
-  handlePayRateClick = () => {
-    this.setState({ showPayRate: true, showTutors: false, showCharts: false });
-  };
-  handleDashboardClick = () => {
-    this.setState({ showTutors: false, showPayRate: false, showCharts: false });
-  };
-  
-
-  componentDidMount = () => {
-    let token = localStorage.getItem('token');
-    if (!token) {
-      this.props.navigate('/login');
-    } else {
-      this.setState({ token: token }, () => {
-        this.getSession();
-
-      });
-    }
+  // Event handler for the register button click
+  handleSessionsClick = () => {
+    this.setState({ showSessions: true, addSession: false, showReports: false });
   };
 
-  getSession = () => {
-    this.setState({ loading: true });
-
-    let data = '?';
-    data = `${data}page=${this.state.page}`;
-    
-    axios
-      .get(`http://localhost:2000/get-users${data}`, {
-        headers: {
-          token: this.state.token,
-        },
-      })
-      .then((res) => {
-        this.setState({ loading: false, users: res.data.users, pages: res.data.pages });
-      })
-      .catch((err) => {
-        this.setState({ loading: false, users: [], pages: 0 });
-      });
+  // Event handler for the accounts button click
+  handleAddSessionClick = () => {
+    this.setState({ showSessions: false, addSession: true, showReports: false });
   };
 
-  pageChange = (e, page) => {
-    this.setState({ page: page }, () => {
-      this.getSession();
-    });
-  };
-
-  handleChartsClick = () => {
-    this.setState({ showTutors: false, showPayRate: false, showCharts: true });
-  };
-
-
-  toggleDrawer = () => {
-    this.setState((prevState) => ({ open: !prevState.open }));
+  // Event handler for the charts button click
+  handleReportsClick = () => {
+    this.setState({ showSessions: false, addSession: false, showReports: true });
   };
 
   LogOutButton = () => {
@@ -210,7 +154,7 @@ class Payroll extends Component {
     const navigate = useNavigate();
   
     const goToWelcomePage = () => {
-      navigate('/WelcomePage');
+      navigate('/tutor/dashboard');
     };
   
     return (
@@ -219,9 +163,9 @@ class Payroll extends Component {
       </IconButton>
     );
   };
-
+  
   render() {
-    const { open, showTutors, showPayRate, showCharts } = this.state;
+    const { open, showSessions, addSession, showReports } = this.state;
     const { LogOutButton, LogoButton } = this;
   
     return (
@@ -246,23 +190,20 @@ class Payroll extends Component {
               >
                 <MenuIcon />
               </IconButton>
-                 <Typography
+              <Typography
                 component="h1"
                 variant="h5"
                 color="white"
                 noWrap
                 sx={{ flexGrow: 1, fontWeight: 'bold', marginLeft: '150px' }} // Adjust this value to your needs
               >
-                Payroll
+                Sessions Dashboard
               </Typography>
               <div>
-              <LogoButton />
-
+                <LogoButton />
                 <span style={{ marginRight: '10px' }}></span> {/* Add space between the two icons */}
-
                 <LogOutButton />
               </div>
-            
             </Toolbar>
           </AppBar>
           <Drawer variant="permanent" open={open}>
@@ -278,19 +219,16 @@ class Payroll extends Component {
                 <ChevronLeftIcon />
               </IconButton>
             </Toolbar>
-            <Divider />
             <List component="nav">
-              {mainListItems(this.handleTutorsClick, this.handlePayRateClick, this.handleDashboardClick, this.handleChartsClick)}
-              <Divider sx={{ my: 1 }} />
+              {/* Render the list Session Options */}
+              {mainListItems(this.handleSessionsClick, this.handleAddSessionClick, this.handleReportsClick)}
             </List>
           </Drawer>
           <Box
             component="main"
             sx={{
               backgroundColor: (theme) =>
-                theme.palette.mode === 'light'
-                  ? theme.palette.grey[100]
-                  : theme.palette.grey[900],
+                theme.palette.mode === 'light' ? theme.palette.grey[100] : theme.palette.grey[900],
               flexGrow: 1,
               height: '100vh',
               overflow: 'auto',
@@ -298,71 +236,23 @@ class Payroll extends Component {
           >
             <Toolbar />
             <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-              {showTutors ? <TutorsList /> : null}
-              {showPayRate ? <PayRate /> : null}
-              {showCharts ? (
-                <iframe 
-                  style={{
-                    background: '#F1F5F4',
-                    border: 'none',
-                    borderRadius: '2px',
-                    boxShadow: '0 2px 10px 0 rgba(70, 76, 79, .2)',
-                    width: '80vw',
-                    height: '100vh'
-                  }}
-                  src="https://charts.mongodb.com/charts-project-0-oyiwi/embed/dashboards?id=c075600e-06f7-4cd2-b9f3-a0bc9ba537c0&theme=light&autoRefresh=true&maxDataAge=3600&showTitleAndDesc=false&scalingWidth=scale&scalingHeight=scale"
-                />
-              ) : null}
-              {!showTutors && !showPayRate && !showCharts ? (
-                <Grid container spacing={3} justify="center">
-                  {/* Recent Deposits */}
-                  <Grid item xs={12} sm={4} md={4} lg={4}>
-                    <Paper
-                      sx={{
-                        p: 2,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        height: 240,
-                      }}
-                    >
-                      <Spendings />
-                    </Paper>
-                  </Grid>
-                  
-                  {/* Recent Deposits */}
-                  <Grid item xs={12} sm={4} md={4} lg={4}>
-                    <Paper
-                      sx={{
-                        p: 2,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        height: 240,
-                      }}
-                    >
-                      <MonthlySpendings />
-                    </Paper>
-                  </Grid>
-                  {/* Recent Deposits */}
-                  <Grid item xs={12} sm={4} md={4} lg={4}>
-                    <Paper
-                      sx={{
-                        p: 2,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        height: 240,
-                      }}
-                    >
-                      <YearlySpendings />
-                    </Paper>
-                  </Grid>
-                  {/* Recent Sessions */}
-                  <Grid item xs={12}>
-                    <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-                      <TutorSnapshot />
-                    </Paper>
-                  </Grid>
-                </Grid>
-              ) : null}
+              {/* Conditionally render the components based on the state */}
+              {showSessions && <Sessions/>}
+              {addSession && <AddSession/>}
+              {showReports && (
+                <iframe
+                style={{
+                  background: '#F1F5F4',
+                  border: 'none',
+                  borderRadius: '2px',
+                  boxShadow: '0 2px 10px 0 rgba(70, 76, 79, .2)',
+                  width: '80vw',
+                  height: '100vh',
+                }}
+                src="https://charts.mongodb.com/charts-project-0-oyiwi/embed/dashboards?id=a6bc5c09-215f-4d3d-942e-fc32036d064f&theme=light&autoRefresh=true&maxDataAge=300&showTitleAndDesc=false&scalingWidth=scale&scalingHeight=scale"
+              />
+              )}
+              {/* Render copyright component */}
               <Copyright sx={{ pt: 4 }} />
             </Container>
           </Box>
@@ -371,9 +261,7 @@ class Payroll extends Component {
     );
   }
   
-
-  
   
 }
 
-export default Payroll;
+export default  (SessionDashbaord);
